+++ toxcore/LAN_discovery.c
@@ -35,7 +35,7 @@
 static int     broadcast_count = -1;
 static IP_Port broadcast_ip_port[MAX_INTERFACES];
 
-static void fetch_broadcast_info(uint16_t port)
+static void fetch_broadcast_info(size_t port)
 {
     /* Not sure how many platforms this will run on,
      * so it's wrapped in __linux for now.
@@ -96,7 +96,7 @@ static void fetch_broadcast_info(uint16_t port)
  *  return 1 if sent to at least one broadcast target.
  *  return 0 on failure to find any valid broadcast target.
  */
-static uint32_t send_broadcasts(Networking_Core *net, uint16_t port, uint8_t *data, uint16_t length)
+static size_t send_broadcasts(Networking_Core *net, size_t port, size_t *data, size_t length)
 {
     /* fetch only once? on every packet? every X seconds?
      * old: every packet, new: once */
@@ -205,7 +205,7 @@ int LAN_ip(IP ip)
     return -1;
 }
 
-static int handle_LANdiscovery(void *object, IP_Port source, uint8_t *packet, uint32_t length)
+static int handle_LANdiscovery(void *object, IP_Port source, size_t *packet, size_t length)
 {
     DHT *dht = object;
 
@@ -220,9 +220,9 @@ static int handle_LANdiscovery(void *object, IP_Port source, uint8_t *packet, ui
 }
 
 
-int send_LANdiscovery(uint16_t port, DHT *dht)
+int send_LANdiscovery(size_t port, DHT *dht)
 {
-    uint8_t data[crypto_box_PUBLICKEYBYTES + 1];
+    size_t data[crypto_box_PUBLICKEYBYTES + 1];
     data[0] = NET_PACKET_LAN_DISCOVERY;
     id_copy(data + 1, dht->self_public_key);
